backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.12.3
boto3==1.42.55
botocore==1.42.55
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
fastapi==0.115.5
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
jmespath==1.1.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.32.1
watchfiles==1.1.1
websockets==16.0
